updated langchain async example
